pyproject.toml
[project]
name = "bluebubbles-linux"
version = "0.1.0"
description = "A GTK4 Linux desktop client for BlueBubbles iMessage server"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Apache-2.0"}
authors = [
    {name = "Nick"}
]
keywords = ["imessage", "bluebubbles", "gtk4", "linux", "hyprland"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: X11 Applications :: GTK",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Chat",
]

dependencies = [
    "PyGObject>=3.46.0",
    "httpx[http2]>=0.27.0",
    "python-socketio[asyncio_client]>=5.11.0",
    "aiohttp>=3.9.0",
    "pydantic>=2.5.0",
    "keyring>=25.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[project.scripts]
bluebubbles = "bluebubbles_linux.main:main"
bluebubbles-panel = "bluebubbles_linux.main:panel"

[project.gui-scripts]
bluebubbles-gui = "bluebubbles_linux.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bluebubbles_linux"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
